lambda/image_processor/requirements.txt
Pillow==10.2.0
boto3==1.34.34 